add muted role init to new service
